New data file for portfolio content
